requirements.txt
# Core
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12

# LLM
openai>=1.0
langchain>=0.2
langchain-openai>=0.1

# Browser automation (for JS-heavy sites — optional)
# selenium>=4.18
# playwright>=1.43

# Scheduling (optional)
# schedule>=1.2
# apscheduler>=3.10
